+++ pricing/src/epilot_pricing/models/address.py
@@ -83,6 +83,7 @@ class Address(BaseModel):
         for n, f in self.model_fields.items():
             k = f.alias or n
             val = serialized.get(k)
+            serialized.pop(k, None)
 
             optional_nullable = k in optional_fields and k in nullable_fields
             is_set = (self.__pydantic_fields_set__.intersection({n}) or k in null_default_fields) # pylint: disable=no-member
@@ -93,6 +94,9 @@ class Address(BaseModel):
                 not k in optional_fields or (optional_nullable and is_set)
             ):
                 m[k] = val
-
+        
+        for k, v in serialized.items():
+            m[k] = v
+        
         return m