+++ pricing/src/epilot_pricing/models/searchstreetsparams.py
@@ -36,6 +36,7 @@ class SearchStreetsParams(BaseModel):
         for n, f in self.model_fields.items():
             k = f.alias or n
             val = serialized.get(k)
+            serialized.pop(k, None)
 
             optional_nullable = k in optional_fields and k in nullable_fields
             is_set = (self.__pydantic_fields_set__.intersection({n}) or k in null_default_fields) # pylint: disable=no-member
@@ -46,6 +47,6 @@ class SearchStreetsParams(BaseModel):
                 not k in optional_fields or (optional_nullable and is_set)
             ):
                 m[k] = val
-
+        
         return m